feat: move search data methods to a separate module
